Center the elements for company page
